requirements.txt
streamlit
tensorflow==2.11.0
numpy
opencv-python-headless
pillow
pandas
scikit-learn
scipy
